chore: move hash_key_value to posthog/models/utils
