+++ .github/workflows/ci-mcp.yml
@@ -328,7 +328,8 @@ jobs:
                   os.environ['DJANGO_SETTINGS_MODULE'] = 'posthog.settings'
                   django.setup()
                   from posthog.models import Organization, Team, User, EventDefinition
-                  from posthog.models.personal_api_key import PersonalAPIKey, hash_key_value
+                  from posthog.models.personal_api_key import PersonalAPIKey
+                  from posthog.models.utils import hash_key_value
                   org = Organization.objects.first()
                   team = Team.objects.first()
                   user = User.objects.first()
