+++ posthog/api/personal_api_key.py
@@ -18,10 +18,10 @@ from posthog.models.activity_logging.personal_api_key_utils import (
     log_personal_api_key_activity,
     log_personal_api_key_scope_change,
 )
-from posthog.models.personal_api_key import LEGACY_HASH_PREFIX, hash_key_value
+from posthog.models.personal_api_key import LEGACY_HASH_PREFIX
 from posthog.models.signals import model_activity_signal, mutable_receiver
 from posthog.models.team.team import Team
-from posthog.models.utils import generate_random_token_personal, mask_key_value
+from posthog.models.utils import generate_random_token_personal, hash_key_value, mask_key_value
 from posthog.permissions import TimeSensitiveActionPermission
 from posthog.scopes import API_SCOPE_ACTIONS, API_SCOPE_OBJECTS
 from posthog.user_permissions import UserPermissions
