+++ posthog/api/test/test_project.py
@@ -8,9 +8,9 @@ from posthog.api.test.test_team import EnvironmentToProjectRewriteClient, team_a
 from posthog.constants import AvailableFeature
 from posthog.models.organization import Organization, OrganizationMembership
 from posthog.models.person import Person
-from posthog.models.personal_api_key import PersonalAPIKey, hash_key_value
+from posthog.models.personal_api_key import PersonalAPIKey
 from posthog.models.project import Project
-from posthog.models.utils import generate_random_token_personal
+from posthog.models.utils import generate_random_token_personal, hash_key_value
 
 
 class TestProjectAPI(team_api_test_factory()):  # type: ignore
