+++ posthog/session_recordings/test/test_session_recording_snapshots_api_blob_v2.py
@@ -7,8 +7,7 @@ from rest_framework import status
 
 from posthog.clickhouse.client import sync_execute
 from posthog.models import Person, PersonalAPIKey, SessionRecording
-from posthog.models.personal_api_key import hash_key_value
-from posthog.models.utils import generate_random_token_personal, uuid7
+from posthog.models.utils import generate_random_token_personal, hash_key_value, uuid7
 from posthog.session_recordings.models.session_recording_event import SessionRecordingViewed
 from posthog.session_recordings.queries.test.session_replay_sql import produce_replay_summary
 from posthog.session_recordings.recordings.errors import RecordingDeletedError
