+++ posthog/test/playwright_setup_functions.py
@@ -14,8 +14,7 @@ from posthog.constants import AvailableFeature
 from posthog.management.commands.generate_demo_data import Command as GenerateDemoDataCommand
 from posthog.models import Dashboard, DashboardTile, Insight, PersonalAPIKey, Team, User
 from posthog.models.insight_variable import InsightVariable
-from posthog.models.personal_api_key import hash_key_value
-from posthog.models.utils import mask_key_value
+from posthog.models.utils import hash_key_value, mask_key_value
 
 
 class PlaywrightSetupVariableType(StrEnum):
