+++ rust/feature-flags/src/api/auth.rs
@@ -240,7 +240,7 @@ async fn load_personal_key_from_pg(
 }
 
 /// Hash a token value using SHA256
-/// Ported from PostHog's `hash_key_value` function in `posthog/models/personal_api_key.py`
+/// Ported from PostHog's `hash_key_value` function in `posthog/models/utils.py`
 /// Used by both PersonalAPIKey and ProjectSecretAPIKey
 pub(crate) fn hash_token_value(value: &str) -> String {
     use sha2::{Digest, Sha256};
